requirements.txt
# 核心依赖
python-dotenv>=1.1.1
lxml>=6.0.2
pillow>=12.0.0
requests>=2.32.5

# 用于更安全/稳健的 XML 解析与美化
defusedxml>=0.7.1

# AI API 支持
zhipuai>=2.1.5
